requirements.txt
py>=1.8.1
pip>=20.0.2
wheel>=0.34.2
pytest>=5.4.1
mypy>=0.770
idna>=2.9
multidict>=4.7.5
Jinja2>=2.11.1
Werkzeug>=1.0.1
itsdangerous>=1.1.0
click>=7.1.1
MarkupSafe>=1.1.1
attrs>=19.3.0
pluggy>=0.13.1
packaging>=20.3
yarl>=1.4.2
aiohttp>=3.6.2
chardet>=3.0.4
wcwidth>=0.1.9
six>=1.14.0
pyparsing>=2.4.7
setuptools>=46.1.3
websockets>=8.1
Flask>=1.1.2

python-socketio>=4.5.1

-e git://github.com/devdoomari3/python-codegen.git@726ff349f8e66018#egg=py_codegen